Project-main-3/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-cas==1.6.0
firebase-admin==6.2.0
sendgrid==6.10.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.3
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
werkzeug==3.0.1

